+++ CHANGELOG.md
@@ -7,7 +7,9 @@ Note: CHANGELOG was created in version 1.1.0, so change logs in versions that pr
 
 ## [unreleased]
 
-No changes.
+### Changed
+
+- Add newfunc and its subfunctions to the class Constructor.
 
 ## [1.1.1] - 2023-9-28