+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Note: CHANGELOG was created in version 1.1.0, so change logs in versions that pr
 
 ## [unreleased]
 
+No changes.
+
+## [1.1.2]
+
 ### Added
 
 - Document strings.
@@ -213,7 +217,8 @@ The first version of MEP released (Versions prior to 0.2.0 are deprecated).
 - ```auto_install.bat``` instead pip.
 - Class ```Formula``` and ```Expression```.
 
-[unreleased]: https://github.com/yuanhang2008/MEP/compare/MEP-v1.1.1...HEAD
+[unreleased]: https://github.com/yuanhang2008/MEP/compare/MEP-v1.1.2...HEAD
+[1.1.2]: https://github.com/yuanhang2008/MEP/compare/MEP-v1.1.1...MEP-v1.1.2
 [1.1.1]: https://github.com/yuanhang2008/MEP/compare/MEP-v1.1.0...MEP-v1.1.1
 [1.1.0]: https://github.com/yuanhang2008/MEP/compare/MEP-v1.0.3...MEP-v1.1.0
 [1.0.3]: https://github.com/yuanhang2008/MEP/compare/MEP-v1.0.2...MEP-v1.0.3